requirements.txt
APScheduler==3.6.1
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
clickclick==1.2.2
colorhash==1.0.2
configparser==4.0.1
connexion==2.3.0
cryptography==2.7
Flask==1.1.1
Flask-MonitoringDashboard==2.1.4
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonpatch==1.24
jsonpointer==2.0
jsonschema==2.6.0
jwcrypto==0.6.0
MarkupSafe==1.1.1
numpy==1.17.2
openapi-spec-validator==0.2.8
Pillow==6.1.0
psutil==5.6.3
pycparser==2.19
PyJWT==1.4.2
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
tzlocal==2.0.0
urllib3==1.25.3
Werkzeug==0.15.5
